loopback: use source sample spec and channel map by default
Currently the loopback module uses sample spec and channel map of the
sink by default. It leads to double resample if source and sink sample
specs are different and no rate/format specified in arguments. This
patch causes the source sample spec and channel map to be used by
default.

%% original patch: 0002-loopback-use-source-sample-spec-and-channel-map-by-d.patch

%% original patch: 0002-loopback-use-source-sample-spec-and-channel-map-by-d.patch